requirements.txt
Django==4.2.6
djangorestframework==3.14.0
drf-yasg==1.21.7
requests==2.31.0
psycopg2
celery
redis
flower
pillow
